+++ src/test/kotlin/com/contentful/java/cma/EntryTests.kt
@@ -129,8 +129,8 @@ class EntryTests : BaseTest() {
     }
 
     @test fun testDelete() {
-        val requestBody = "203"
-        server!!.enqueue(MockResponse().setResponseCode(200).setBody(requestBody))
+        val requestBody = ""
+        server!!.enqueue(MockResponse().setResponseCode(204).setBody(requestBody))
         assertTestCallback(client!!.entries().async().delete(
                 "spaceid", "entryid", TestCallback()) as TestCallback)